Turn some stairs in Tomb into escape hatches with fixed destinations
To prevent stair dancing from being such a pervasive technique for Tomb
levels, we replace bidirectional stone stairs with escape hatches that
have a fixed destination. This is an initial implementation, so the
escape locations back up on Tomb:2 and Tomb:3 haven't been given much
thought.

The first staircase from Tomb:1 to Tomb:2 is now a hatch landing where
the Tomb:2 ambush stairs used to be.

For Tomb:2, the stone stairs going back to Tomb:1 are still stone, which
should probably be change, since players will be very compelled to try
to teleport into the Tomb:1 hallway.

To get to Tomb:3, there are two escape hatches on Tomb:2 leading to the
same spot on Tomb:3 where there used to be three stone upstairs. Escape
hatches leading back to the Tomb:2 outer rooms are in each treasure
chamber.