Save compatability for the change of stone stairs into hatches in Tomb
Only needed in a game that has generated a Tomb level already with stone
stairs but has at least one ungenerated Tomb level. Since the already
generated level will have stone stairs and no necessary hatch name
markers, we generate the remaining vaults with stone stairs using 4
specially tagged vaults in tomb.des.

If such a game is loaded, a property is set in you.props that will have
the builder load the necessary vault so that stone stairs are generated
throughout. This level is only ever one specific layout, but should
otherwise be a normal level. Old games that haven't generated any Tomb
levels will get hatches normally, and old games that have already
generated all Tomb levels are unaffected.

These vaults will have to be maintained to a reasonable degree before
TAG_MAJOR is incremented, which is one disadvantage to this approach.
However cleaning up the level to make the hatches upon entry would
(probably) be more difficulty and prone to breakage.

Tested with saves from trunk with and without a generated Tomb:1, and
the compat code appears to work fine.